Add files via upload
Question1:
Create an inline function to calculate the area of a sphere.
In your main function take the radius from the user and store these values inside a double
array. Keep storing values so long as the user input isn’t a negative value. Then call your
inline function to calculate the area for each of the sphere and print them onto the screen.